:recycle: refactor: move auth-related classes to auth domain package
                - GoogleLoginController → auth.controller로 이동
                - 인증 응답 DTO (GoogleOAuthResponse, GoogleTokenResponse 등) → auth.dto로 이동
                - AuthService, GoogleOAuthService → auth.service 및 impl로 이동
                - SecurityConfig, 필터, 기존 서비스 코드에서 import 경로 수정
                - 인증 도메인을 분리하여 구조 명확화 및 유지보수성 향상
