+++ src/main/java/com/handongapp/cms/auth/service/GoogleOAuthService.java
@@ -1,6 +1,6 @@
-package com.handongapp.cms.service;
+package com.handongapp.cms.auth.service;
 
-import com.handongapp.cms.security.dto.GoogleOAuthResponse;
+import com.handongapp.cms.auth.dto.GoogleOAuthResponse;
 
 public interface GoogleOAuthService {
     GoogleOAuthResponse authenticate(String authorizationCode);
