+++ src/main/java/com/handongapp/cms/service/TbUserService.java
@@ -1,7 +1,7 @@
 package com.handongapp.cms.service;
 
 import com.handongapp.cms.domain.TbUser;
-import com.handongapp.cms.security.dto.GoogleUserInfoResponse;
+import com.handongapp.cms.auth.dto.GoogleUserInfoResponse;
 
 public interface TbUserService {
     TbUser saveOrUpdateUser(String userId, String email, String name);
